requirements.txt
python-dateutil
flask~=3.0.2
RPi.GPIO
RPi.GPIO-def